made the anchor element of grid item to coloum
This means that even if the grid item isn't inside a grid it will still stack vertically.

I think in the future I should make all of these individual and reusable classes but for now I'll do it like this so I can get an idea of where I went wrong with hindsight.