testing for playlist button
gonna abandon this branch for now. using the YouTube API requires a key and I can't obscure it using java script since this repository is required to be public so I'll learn to set up backend using google cloud so that I can create  an email forum and put my API key behind it.