created a svg alternative to profile pictures
this could be expanded to include icons for roles, like programmer, artist, sfx, etc. could be a fun way to do the profile pictures so that I don't have to host a bunch of images, or so that I don't have to hunt down a profile picture for each collaborator.

I'll have to think on this problem.