moved styles to the web component file.
This means that it won't need to make another request, and items shouldn't have any down time of not being styled.

also makes editing the web component simpler since it can all be done in one place.
File size being longer is a worry, but everything is small enough that it should be easy to find what you are looking for.